audio: change ao_get_delay() from seconds to samples
The whole purpose of this change is to free the code of GPL-only changes
done in 2001. It partially reverts commits 242aa6ebd and 4943e9c52. The
author, anders, disagreed with relicensing to LGPL.

These commits changed the per-driver get_delay callback to return the
delay in seconds instead of audio bytes. This new commit changes the
callback from seconds to samples, which should remove any residual
copyright as much as the original change introduced it. The other
changes (creating the ao_data_t struct) are not considered by this
commit.