audio/out: change license of some core files to LGPL
Most contributors have agreed. ao.c requires a "driver" entry for each
audio output - we assume that if someone who didn't agree to LGPL added
a line, it's fine for ao.c to be LGPL anyway. If the affected audio
output is not disabled at compilation time, the resulting binary will be
GPL anyway.

A very big problem is commit 242aa6ebd406 by anders, who disagreed with
the LGPL. This commit adds an "AO plugin" framework, which was moved to
libaf (now audio/filter/) later - this AO plugin thing is completely
gone from the AO code. But the commit also changes the proper AO
interface, and does two things:

1. It changes the get_delay function.
2. It moves some global variables (like ao_sampledata) into a struct.

We consider 1. solved by the previous commit.

For 2., the situation is complicated. The "struct ao_data_s" was renamed
to "struct ao" later on, and the "samplerate", "format", and "channels"
fields still exist (the other fields were removed, see 4d3a2c7e0dac3 for
an example). On the other hand, the ao_data variable is gone and was
replaced by a pointer. It could be argued that changing all ao_data uses
is as big of a change as anders' original change, and overwrites them
all. So the remaining code relevant for copyright are only these 3
struct fields. Since moving global variables to struct fields, and
driver-like frameworks (like VO) moving to have context structs (like
struct ao) has been a very common thing in the MPlayer/mpv codebase, it
could be argued that anders' copyright does not affect this code
anymore.

The change to pass "struct ao" as context to audio outputs was done in
3a5fd15fa2568, and then incrementally for each AO. 40f6ab5064a62 is the
first commit that updates an AO with the change.

The audio output code itself was unspired or partically copied from
libao in 7a2eec4b59f4 (thus why MPlayer's audio code is named libao2).
Just to be sure we got permission from Jack Moffitt and Stan Seibert,
who according to libao's SVN history are the initial author. Aaron
Holtzman is also an initial author - we did not get explicit permission
from him, but assume it's fine, as we got granted permission for the
VO/libvo parts (as far as "residual copyright" of his work in mpv goes).
We have to assume this, as he was hard to reach, and we forgot to ask
about libao at first.

d65c8518dee: unknown whether copyrightable, but awaiting permission from
Jens Axboe.

0ef8e555735: nick could not be reached, but the include statement was
removed again anyway.

879e05a7c17: iive agreed to LGPL v3+ only, but this line of code was
removed anyway, so ao_null.c can be LGPL v2.1+.

9dd8f241ac2: patch author could not be reached, but the corresponding
code (old slave mode interface) was completely removed later.